requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
amqp==5.3.1
    # via kombu
asgiref==3.10.0
    # via
    #   channels
    #   channels-redis
    #   django
    #   django-cors-headers
billiard==4.2.2
pyJianYingDraft
opencv-python
    # via celery
celery==5.5.0b2
    # via
    #   ai-story (pyproject.toml)
    #   django-celery-beat
    #   django-celery-results
channels==4.0.0
    # via
    #   ai-story (pyproject.toml)
    #   channels-redis
channels-redis==4.2.1
    # via ai-story (pyproject.toml)
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
django==3.2.15
    # via
    #   ai-story (pyproject.toml)
    #   channels
    #   django-celery-beat
    #   django-cors-headers
    #   django-filter
    #   django-redis
    #   django-timezone-field
    #   djangorestframework
django-celery-beat==2.2.1
    # via ai-story (pyproject.toml)
django-celery-results==2.4.0
    # via ai-story (pyproject.toml)
django-cors-headers==4.5.0
    # via ai-story (pyproject.toml)
django-filter==23.5
    # via ai-story (pyproject.toml)
django-redis==5.4.0
    # via ai-story (pyproject.toml)
django-timezone-field==4.2.3
    # via django-celery-beat
djangorestframework==3.14.0
    # via ai-story (pyproject.toml)
kombu==5.5.4
    # via celery
msgpack==1.1.2
    # via channels-redis
packaging==25.0
    # via kombu
prompt-toolkit==3.0.52
    # via click-repl
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via celery
pytz==2025.2
    # via
    #   django
    #   django-timezone-field
    #   djangorestframework
redis==6.4.0
    # via
    #   channels-redis
    #   django-redis
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via
    #   celery
    #   kombu
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
djangorestframework-simplejwt